implement a very rigged way handle searching for Conditions by code
This method probably isn't sufficient because the lookup logic now involves
code, e.g matches for condition code involves a filter. This will be hard
to encode when using the searchparams to create the logic.

Its likely that using a datalog query instead of a datoms lookup will be the
easiest way to handle this.