enable starting server from dev ns
currently requires you to specify the database uri in the blaze.edn,
though maybe there is the option to send it in through env vars.

this removes datomic free from the module because if both free and pro are found
it might lead to conflicts. The solution here might be to exclude datomic-free
when including datomic pro. Which might be appropriate on a per use basis. As
in, when releasing, when connecting to a pro database. That would probably
mean the top level deps and the module deps do include default datomic free
clients. Then either build a run tool that users can take advantage of, or
document the deployment override strategy so users know how to take advantage
of it.