pull type from URI.
Previously the type was pulled from the path param. Sense the type is no longer
a parameter this wasnt working. Pulling it from the URI isn't right either,
as the URI isn't right either. If the handlers are on a per type basis then the
same place in the configuration where we specify the type is where the
type probably should be passed to the handler. Given how things are currently
structured its not clear the best way to do that.