use working dev namespace
directly counteracting my last commit, the dev/blaze/dev file/namespace
does work but it just needed the clojure tools lib. This is loaded with the
clj "test" alias but that also directly rush a test runner and so cant be used
for development. Its a mystery to me how this works as is, but one solution is
to just create and use a dev namesapce that contains the missing lib.