Currently we are placing the transformations before the expression
filters. The transformations seem to prep data for the rest api response. So
they aren't namespaced e.g code vs fhir.Coding/code.

This raises the question of how the expressions should be... expressed.
Should they map whats in the db or map the fhir spec were trying to present
in the API?

Using code here for the moment delays having to update the transformation
Process if we choose the later.