pull Medication data that matches existing data.
pull medication name instead of code$cr. This has human readable info. The
other was a code that i assume we could use to look it up. But not sure
why that would be better.

pull prescriber name text over display. None of john does records had
data for display.