Use $cr attribute.
we should rely on $cr attributes for performance and then the transformation
destructures them which is why are search params expression. The destructuring
doesnt produce namespaces hense why its just :code.

Before we used the pattern with text it was category$cr which doesn't seem
right it needs to be category [:fhir.CodeableConcept/coding$cr]. This
was why it wasn't working originally.