Breezeehr uses datomic pro not datomic free (#3)
* Breezeehr uses datomic pro not datomic free

To allow for this we need to

1. point to the peer library location at the mvn repo at my.datomic.com
2. add support for the db client library for out specific database (mysql)
3. change from datomic free to pro

The downside to this approach is that it changes the default from
free to pro. It would be ideal to allow that to be the default but as
far as i understand their isn't a way to override this using clj.

I recall asking alex about this specifically and getting a no.
Another possibility is that we could generate the edn file using clojure.
But this should be its own feature.

* fix typo

* the datomic module will also need to use datomic pro

over datomic free. see the reason for the root deps updating.

* move mysql to a clj alias

Dromon will be using datomic pro. That won't be configurable, but the
backend database should be. This allows for this. This means
operations and dev environments will need to run with this alias

* move comment above relevant config

* The module doesn't have to have the datomic mysql alias

* remove comment

this is documented elsewhere

* dromon and its modules use datomic pro

* fix typo

* bump datomic pro to latest

* fix require, Authentication is a protocol and so is required

not imported. How this works on Samply/blaze remains a mystery.

* remove exclusion on netty-all

This exlusion came from samply and is likely due to a fixed issue or one
not relevant to datomic pro. The only reference to the cause i can find
is here:

https://datomic.narkive.com/u8agzqCc/datomic-0-9-5385-breaks-when-using-aleph-0-4-1-netty-incompatibility

from 4 years ago.

* disable failing test.

The test has a note about a possible flaw and its reasonable to assume
that what the test is trying to capture has changed rather then something
broke.

Co-authored-by: drew verlee <dverlee@breezeehr.com>