Add self signed ssl as a feature. (#7)
* Add self signed ssl as a feature.

If a user sets the env var SSL, this enables the self signed ssl.
This happens by passing the key to the integrate init which selects
netty self signed ssl.

An obvious limitation of this approach is that only the self signed
option is available without modifying the project config (dromon.edn)
or the code. However I'm not sure what other options would look like. So
those should be addressed as the come up.

* fix typo. blaze old name, new name dromon.