make server init open for extension.
Previously it took a set number of arguments, however, depending on the
context, the way it used those arguments changed. e.g for metrics
you dont need to use a self signed cert.

Complicating this issue is that it was speced. Here again, a spec depends on
its context.

By passing a map we allow the init function to ignore arguments it doesn't
care about, while passing on those it does.