move dromon edn out of artifact when its built.
the src class path contains everything that should be built with
the project. previously we included dromon.edn in that, but moving
forward it will be more a more dynamic build time artifact. Moving it
to the dev class path allows it to exist as a useful
development/default/starting config.